Move more constructors to device
