+++ src/element-buffer.ts
@@ -1,8 +1,7 @@
 import * as assert from "./util/assert";
-import * as array from "./util/array";
-import { BufferUsage, DataType, Primitive, sizeOf } from "./types";
+import { BufferUsage, DataType, Primitive } from "./types";
 
-export type Device = import ("./device").Device;
+export type Device = import("./device").Device;
 
 export type ElementArray =
     | number[] // infers POINTS
@@ -53,85 +52,6 @@ export interface ElementBufferStoreOptions {
  */
 export class ElementBuffer<T extends ElementBufferType> {
 
-    /**
-     * Create a new element buffer with given type, primitive, and size.
-     */
-    static create<T extends ElementBufferType>(
-        dev: Device,
-        type: T,
-        primitive: Primitive,
-        size: number,
-        { usage = BufferUsage.DYNAMIC_DRAW }: ElementBufferOptions = {},
-    ): ElementBuffer<T> {
-        return new ElementBuffer(
-            dev._gl,
-            type,
-            primitive,
-            size,
-            size * sizeOf(type),
-            usage,
-        );
-    }
-
-    /**
-     * Create a new element buffer from potentially nested array. Infers
-     * Primitive from the array's shape:
-     *   number[] -> POINTS
-     *   [number, number][] -> LINES
-     *   [number, number, number][] -> TRIANGLES
-     * Does not take ownership of data.
-     */
-    static withArray(
-        dev: Device,
-        data: ElementArray,
-        options?: ElementBufferOptions,
-    ): ElementBuffer<DataType.UNSIGNED_INT> {
-        if (array.is2(data)) {
-            const shape = array.shape2(data);
-            assert.rangeInclusive(shape[1], 2, 3, (p) => {
-                return `Elements must be 2-tuples or 3-tuples, got ${p}-tuple`;
-            });
-            const ravel = array.ravel2(data, shape);
-            const primitive = shape[1] === 3
-                ? Primitive.TRIANGLES
-                : Primitive.LINES;
-            return ElementBuffer.withTypedArray(
-                dev,
-                DataType.UNSIGNED_INT,
-                primitive,
-                ravel,
-            );
-        }
-        return ElementBuffer.withTypedArray(
-            dev,
-            DataType.UNSIGNED_INT,
-            Primitive.POINTS,
-            data,
-            options,
-        );
-    }
-
-    /**
-     * Create a new element buffer of given type with provided data. Does not
-     * take ownership of data.
-     */
-    static withTypedArray<T extends ElementBufferType>(
-        dev: Device,
-        type: T,
-        primitive: Primitive,
-        data: ElementBufferTypeToTypedArray[T] | number[],
-        { usage = BufferUsage.STATIC_DRAW }: ElementBufferOptions = {},
-    ): ElementBuffer<T> {
-        return new ElementBuffer(
-            dev._gl,
-            type,
-            primitive,
-            data.length,
-            data.length * sizeOf(type),
-            usage,
-        ).store(data);
-    }
-
     readonly type: T;
     readonly length: number;
     readonly byteLength: number;
@@ -142,7 +62,7 @@ export class ElementBuffer<T extends ElementBufferType> {
 
     private gl: WebGL2RenderingContext;
 
-    private constructor(
+    constructor(
         gl: WebGL2RenderingContext,
         type: T,
         primitive: Primitive,
